+++ app/services/reports/questions_csv_report_service.rb
@@ -1,29 +1,6 @@
 require "csv"
 
-class Reports::CsvReportsService
-  FORM_CSV_HEADERS = [
-    "Form ID",
-    "Status",
-    "Form name",
-    "Slug",
-    "Organisation name",
-    "Organisation ID",
-    "Group name",
-    "Group ID",
-    "Created at",
-    "Updated at",
-    "Number of questions",
-    "Has routes",
-    "Payment URL",
-    "Support URL",
-    "Support URL text",
-    "Support email",
-    "Support phone",
-    "Privacy policy URL",
-    "What happens next markdown",
-    "Submission type",
-  ].freeze
-
+class Reports::QuestionsCsvReportService
   IS_REPEATABLE = "Is repeatable?".freeze
   QUESTIONS_CSV_HEADERS = [
     "Form ID",
@@ -57,16 +34,6 @@ class Reports::CsvReportsService
     @form_documents = form_documents
   end
 
-  def forms_csv
-    CSV.generate do |csv|
-      csv << FORM_CSV_HEADERS
-
-      form_documents.each do |form_document|
-        csv << form_row(form_document)
-      end
-    end
-  end
-
   def questions_csv(answer_type: nil)
     CSV.generate do |csv|
       csv << QUESTIONS_CSV_HEADERS
@@ -96,33 +63,6 @@ class Reports::CsvReportsService
   end
 
 private
-
-  def form_row(form)
-    form_id = form["form_id"]
-    group = GroupForm.find_by_form_id(form_id)&.group
-    [
-      form_id,
-      form["tag"],
-      form["content"]["name"],
-      form["content"]["form_slug"],
-      group&.organisation&.name,
-      group&.organisation&.id,
-      group&.name,
-      group&.external_id,
-      form["content"]["created_at"],
-      form["content"]["updated_at"],
-      form["content"]["steps"].length,
-      form["content"]["steps"].any? { |step| step["routing_conditions"].present? },
-      form["content"]["payment_url"],
-      form["content"]["support_url"],
-      form["content"]["support_url_text"],
-      form["content"]["support_email"],
-      form["content"]["support_phone"],
-      form["content"]["privacy_policy_url"],
-      form["content"]["what_happens_next_markdown"],
-      form["content"]["submission_type"],
-    ]
-  end
 
   def question_rows(form, answer_type)
     form_id = form["form_id"]
